move categories to own file
