+++ src/components/common/categories.js
@@ -1,0 +1,9 @@
+const categories = [
+  'Professional time',
+  'Obligatory time',
+  'Personal time',
+  'Self time',
+  'Non-productive time',
+]
+
+export default categories
